vizard_private_project/backend/requirements.txt
Flask
yt-dlp
ffmpeg-python
pydub
openai
numpy
